+++ src/main/java/com/tenpo/adder/addition/service/AdditionService.java
@@ -2,6 +2,6 @@ package com.tenpo.adder.addition.service;
 
 public interface AdditionService {
 
-    Integer calculateAddition(Integer numberOne, Integer numberTwo);
+    Double calculateAddition(Double numberOne, Double numberTwo);
 
 }
